Develop business logic: update models with necessary fields, update mappers, initialize simple validation and exception handler

+++ src/main/java/lv/klix/oas/controller/OffersResponse.java
@@ -18,7 +18,7 @@ public class OffersResponse {
         private UUID id;
         private BigDecimal monthlyPaymentAmount;
         private BigDecimal totalRepaymentAmount;
-        private BigDecimal numberOfPayments;
+        private Integer numberOfPayments;
         private BigDecimal annualPercentageRate;
         private LocalDate firstRepaymentDate;
     }
